src/test/java/learn/springframework/controllers/ConstructorInjectedControllerTest.java
package learn.springframework.controllers;

import static org.junit.Assert.assertEquals;

import org.junit.Before;
import org.junit.Test;

import learn.springframework.services.GreetingServiceImpl;

public class ConstructorInjectedControllerTest {

	private ConstructorInjectedController constructorInjectedController;

	@Before
	public void setUp() throws Exception {
		this.constructorInjectedController = new ConstructorInjectedController(new GreetingServiceImpl());
	}

	@Test
	public void testSayHello() {
		assertEquals(GreetingServiceImpl.SAY_HELLO, constructorInjectedController.sayHello());
	}

}
